refactor: Printer클래스의 역할 제한 및 관련 메소드 이동
